.archive-7.22.25/agents/vana_v1_archived_2025_01_22/vana_active/requirements.txt
google-adk>=1.1.1
google-cloud-aiplatform>=1.95.1
google-genai>=1.18.0
pydantic>=2.11.5
pydantic-settings>=2.9.1
python-dotenv>=1.1.0
aiohttp>=3.9.0
httpx>=0.28.1
requests>=2.32.3
pyyaml>=6.0.2
psutil>=7.0.0
google-generativeai>=0.8.5
pytz>=2025.2